override http_port and probe target port is ssl is enabled"
chore:  Update the port in the probe when TLS is enabled.
        This propagates the configuration to the transformer pod as well
        override the ports configuration.
