+++ pkg/constants/constants_odh.go
@@ -67,25 +67,15 @@ const (
 // ODH Connections API
 const ODHS3Endpoint = "AWS_S3_ENDPOINT"
 
-// TLS infrastructure for service-ca bundle mounting and transformer-to-predictor communication
+// TLS infrastructure for service-ca bundle mounting and transformer-to-predictor communication (OpenShift-specific)
 const (
 	ServiceCaBundleVolumeName = "openshift-service-ca-bundle"
 	ServiceCaBundleMountPath  = "/etc/odh/openshift-service-ca-bundle"
 	ServiceCaBundleCertFile   = "service-ca.crt"
 
-	// Env var keys for predictor TLS endpoint discovery
-	PredictorHostEnvVar     = "PREDICTOR_HOST"
-	PredictorPortEnvVar     = "PREDICTOR_PORT"
-	PredictorProtocolEnvVar = "PREDICTOR_PROTOCOL"
-
-	// CLI arg to enable SSL for transformer→predictor communication
-	ArgumentPredictorUseSSL = "--predictor_use_ssl"
-
-	// Transformer serving-cert volume/mount/env for native TLS (HTTPS on 8443)
+	// Transformer serving-cert volume/mount for native TLS (HTTPS on 8443)
 	TransformerTLSVolumeName = "transformer-tls"
 	TransformerTLSMountPath  = "/etc/tls/private"
-	TransformerTLSCertEnvVar = "KSERVE_TLS_CERT_FILE"
-	TransformerTLSKeyEnvVar  = "KSERVE_TLS_KEY_FILE"
 	TransformerHTTPSPort     = int32(8443)
 )
